--- src/main/java/ifpb/User.java
@@ -1,8 +1,0 @@
-package ifpb;
-
-public interface User {
-    Name getName();
-    void setName(Name name);
-    Email getEmail();
-    void setEmail(Email email);
-}
