+++ src/main/java/ifpb/collections/interfaces/IAuthors.java
@@ -4,9 +4,13 @@ import ifpb.entitybasic.interfaces.IAuthor;
 import ifpb.entitybasic.interfaces.IID;
 
 public interface IAuthors {
-    void add(IAuthor author);
-    void remove(IID id);
-    IAuthors list();
-    void update(IID id, IAuthor author);
-    IAuthor get(IAuthor author);
+
+        void add(IAuthor author);
+        void remove(String matricula);
+        IAuthor get(IID matricula);
+        IAuthor[] get();
+
+        void update(IAuthor author);
+
 }
+
