+++ src/main/java/ifpb/entitybasic/Name.java
@@ -3,6 +3,8 @@ package ifpb.entitybasic;
 import ifpb.entitybasic.exceptions.InvalidNameException;
 import ifpb.entitybasic.interfaces.IName;
 
+import java.util.Arrays;
+
 public class Name implements IName {
     private String name;
 
@@ -14,11 +16,19 @@ public class Name implements IName {
     public void changeName(String newName) {
         this.name = validateName(newName);
     }
+    @Override
+    public String getName() {
+        return name;
+    }
 
     private String validateName(String name){
         if (name == null || name.isEmpty()){
             throw new InvalidNameException();
         }
-        return name;
+        return formatName(name);
+    }
+
+    private String formatName(String name){
+        return Arrays.stream(name.split(" ")).map((word -> word.substring(0,1).toUpperCase().concat(word.substring(1).toLowerCase()))).toString();
     }
 }
