+++ src/main/java/ifpb/entitybasic/interfaces/IAuthor.java
@@ -3,7 +3,6 @@ package ifpb.entitybasic.interfaces;
 import ifpb.entitybasic.Author;
 
 public interface IAuthor extends IBasicEntity<IAuthor> {
-    boolean isEmpty();
     boolean compareKeys(IID id);
     String getName();
     IID<String> getId();
