+++ src/main/java/ifpb/entitybasic/interfaces/IBasicEntity.java
@@ -2,6 +2,4 @@ package ifpb.entitybasic.interfaces;
 
 public interface IBasicEntity<T> {
     boolean compareTo(T compareData);
-    boolean isEmpty();
-
 }
