+++ src/main/java/ifpb/entitybasic/interfaces/IDescription.java
@@ -1,0 +1,6 @@
+package ifpb.entitybasic.interfaces;
+
+public interface IDescription {
+    void changeDescription(String description) throws Exception;
+    String getDescription();
+}
