+++ src/main/java/ifpb/entitybasic/interfaces/IKeyWord.java
@@ -4,5 +4,4 @@ import ifpb.entitybasic.KeyWord;
 
 public interface IKeyWord extends IBasicEntity<KeyWord> {
     boolean compareTo( KeyWord compareData);
-    boolean isEmpty();
 }
