+++ src/main/java/ifpb/entitybasic/interfaces/IName.java
@@ -2,4 +2,5 @@ package ifpb.entitybasic.interfaces;
 
 public interface IName {
     void changeName(String newName);
+    String getName();
 }
