+++ src/main/java/ifpb/entitybasic/interfaces/ITitle.java
@@ -1,0 +1,6 @@
+package ifpb.entitybasic.interfaces;
+
+public interface ITitle {
+    void changeTitle(String title) throws Exception;
+    String getTitle();
+}
